Simplify synchronization manager
